+++ batman/functions/analytical.py
@@ -58,7 +58,7 @@ import numpy as np
 from .utils import multi_eval
 
 
-class SixHumpCamel(object):
+class SixHumpCamel:
     r"""SixHumpCamel class [Molga2005]_.
 
     .. math:: \left(4-2.1x_1^2+\frac{x_1^4}{3}\right)x_1^2+x_1x_2+
@@ -95,7 +95,7 @@ class SixHumpCamel(object):
         return f
 
 
-class Branin(object):
+class Branin:
     r"""Branin class [Forrester2008]_.
 
     .. math:: f(x) = \left( x_2 - \frac{5.1}{4\pi^2}x_1^2 + \frac{5}{\pi}x_1 - 6
@@ -134,7 +134,7 @@ class Branin(object):
         return f
 
 
-class Michalewicz(object):
+class Michalewicz:
     r"""Michalewicz class [Molga2005]_.
 
     It is a multimodal *d*-dimensional function which has :math:`d!`
@@ -179,7 +179,7 @@ class Michalewicz(object):
         return -f
 
 
-class Rosenbrock(object):
+class Rosenbrock:
     r"""Rosenbrock class [Dixon1978]_.
 
     .. math:: f(x)=\sum_{i=1}^{d-1}[100(x_{i+1}-x_i^2)^2+(x_i-1)^2]
@@ -218,7 +218,7 @@ class Rosenbrock(object):
         return f
 
 
-class Rastrigin(object):
+class Rastrigin:
     r"""Rastrigin class [Molga2005]_.
 
     It is a multimodal *d*-dimensional function which has regularly distributed
@@ -257,7 +257,7 @@ class Rastrigin(object):
         return f
 
 
-class Ishigami(object):
+class Ishigami:
     r"""Ishigami class [Ishigami1990]_.
 
     .. math:: F = \sin(x_1)+7\sin(x_2)^2+0.1x_3^4\sin(x_1), x\in [-\pi, \pi]^3
@@ -311,7 +311,7 @@ class Ishigami(object):
         return f
 
 
-class G_Function(object):
+class G_Function:
     r"""G_Function class [Saltelli2000]_.
 
     .. math:: F = \Pi_{i=1}^d \frac{\lvert 4x_i - 2\rvert + a_i}{1 + a_i}
@@ -361,7 +361,7 @@ class G_Function(object):
         return f
 
 
-class Forrester(object):
+class Forrester:
     r"""Forrester class [Forrester2007]_.
 
     .. math:: F_{e}(x) = (6x-2)^2\sin(12x-4), \\
@@ -407,7 +407,7 @@ class Forrester(object):
         return f
 
 
-class Channel_Flow(object):
+class Channel_Flow:
     r"""Channel Flow class.
 
     .. math:: \frac{dh}{ds}=\mathcal{F}(h)=I\frac{1-(h/h_n)^{-10/3}}{1-(h/h_c)^{-3}}\\
@@ -446,11 +446,11 @@ class Channel_Flow(object):
                          "width={}".format(dx, length, width))
 
     @multi_eval
-    def __call__(self, x, full=True):
+    def __call__(self, x, h_nc=False):
         """Call function.
 
         :param list x: inputs [Ks, Q].
-        :param bool full: Whether to return hc and hn.
+        :param bool h_nc: Whether to return hc and hn.
         :return: Water height along the channel.
         :rtype: array_like (n_samples, n_features [+ 2])
         """
@@ -466,10 +466,10 @@ class Channel_Flow(object):
 
         z_h = self.zref + h
 
-        return np.append(z_h, np.array([hc, hn])) if full else z_h
+        return np.append(z_h, np.array([hc, hn])) if h_nc else z_h
 
 
-class Manning(object):
+class Manning:
     """Manning equation for rectangular channel class."""
 
     logger = logging.getLogger(__name__)
@@ -512,7 +512,7 @@ class Manning(object):
         return h
 
 
-class ChemicalSpill(object):
+class ChemicalSpill:
     r"""Environmental Model class [Bliznyuk2008]_.
 
     Model a pollutant spill caused by a chemical accident.